agt-payment/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0
django-redis==5.4.0
redis==5.0.4
PyJWT==2.8.0
cryptography==42.0.5
drf-spectacular==0.27.2
python-decouple==3.8
httpx==0.27.0
gunicorn==22.0.0
pytest==8.2.0
pytest-django==4.8.0
coverage==7.5.1
